requirements.txt
flask==3.0.2
google-generativeai==0.3.2
python-dotenv==1.0.1
gunicorn==21.2.0
Werkzeug==3.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7 